requirements.txt
aiogram
pillow
torch
torchvision